Major Update

+++ app/Http/Controllers/LaporanController.php
@@ -5,8 +5,11 @@ namespace App\Http\Controllers;
 use DB;
 use App\Laporan;
 use App\Riwayat;
+use App\Exports\LaporanExport;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Maatwebsite\Excel\Facades\Excel;
+use Symfony\Component\HttpFoundation\BinaryFileResponse;
 
 class LaporanController extends Controller
 {
@@ -37,13 +40,59 @@ class LaporanController extends Controller
             'keywordTahun' =>$keywordTahun,
         ];
 
-        if ($user->role == 4){
+        if ($user->role == 1){
+            $result = Laporan::when($request->keywordDaerah != null, function ($query) use ($request) {
+                    $query->where('lap_domisili', $request->keywordDaerah)
+                ;})
+                ->when($request->keywordBulan != null, function ($query) use ($request) {
+                    $query->whereMonth('lap_tanggal', $request->keywordBulan)
+                ;})
+                ->when($request->keywordTahun != null, function ($query) use ($request) {
+                    $query->whereYear('lap_tanggal', $request->keywordTahun)
+                ;})
+                ->get();
+
+            $result2 = DB::table('laporan')
+                ->select(DB::raw('lap_domisili as area, YEAR(lap_tanggal) as year, MONTH(lap_tanggal) as month'), DB::raw('sum(lap_jml) as total'))
+                ->groupBy('area', 'year', 'month')
+                ->when($request->keywordDaerah != null, function ($query) use ($request) {
+                    $query->where('lap_domisili', $request->keywordDaerah)
+                ;})
+                ->when($request->keywordBulan != null, function ($query) use ($request) {
+                    $query->whereMonth('lap_tanggal', $request->keywordBulan)
+                ;})
+                ->when($request->keywordTahun != null, function ($query) use ($request) {
+                    $query->whereYear('lap_tanggal', $request->keywordTahun)
+                ;})
+                ->orderBy('lap_tanggal', 'desc')
+                ->simplePaginate(10);
+        }
+        elseif ($user->role == 3){
+            $result = Laporan::where('lap_domisili', $user->domisili)
+                ->when($request->keywordBulan != null, function ($query) use ($request) {
+                    $query->whereMonth('lap_tanggal', $request->keywordBulan)
+                ;})
+                ->when($request->keywordTahun != null, function ($query) use ($request) {
+                    $query->whereYear('lap_tanggal', $request->keywordTahun)
+                ;})
+                ->get();
+
+            $result2 = Laporan::where('lap_domisili', $user->domisili)
+                ->when($request->keywordBulan != null, function ($query) use ($request) {
+                    $query->whereMonth('lap_tanggal', $request->keywordBulan)
+                ;})
+                ->when($request->keywordTahun != null, function ($query) use ($request) {
+                    $query->whereYear('lap_tanggal', $request->keywordTahun)
+                ;})
+                ->orderBy('lap_tanggal', 'desc')
+                ->simplePaginate(10);
+        }
+        elseif ($user->role == 4){
             $result = DB::table('riwayat_donasi')
                 ->join('donatur', 'donatur.dona_id', '=', 'riwayat_donasi.user_id')
-                ->join('users', 'users.id', '=', 'riwayat_donasi.fund_id')
-                ->where('riwayat_donasi.fund_id', $user->id)
+                ->where('donatur.fund_id', $user->id)
                 ->when($request->keywordDaerah != null, function ($query) use ($request) {
-                    $query->where('domisili', $user->domisili)
+                    $query->where('dona_kota_kab', $request->keywordDaerah)
                 ;})
                 ->when($request->keywordBulan != null, function ($query) use ($request) {
                     $query->whereMonth('riwa_tanggal', $request->keywordBulan)
@@ -55,10 +104,9 @@ class LaporanController extends Controller
 
             $result2 = DB::table('riwayat_donasi')
                 ->join('donatur', 'donatur.dona_id', '=', 'riwayat_donasi.user_id')
-                ->join('users', 'users.id', '=', 'riwayat_donasi.fund_id')
-                ->where('riwayat_donasi.fund_id', $user->id)
+                ->where('donatur.fund_id', $user->id)
                 ->when($request->keywordDaerah != null, function ($query) use ($request) {
-                    $query->where('domisili', $user->domisili)
+                    $query->where('dona_kota_kab', $request->keywordDaerah)
                 ;})
                 ->when($request->keywordBulan != null, function ($query) use ($request) {
                     $query->whereMonth('riwa_tanggal', $request->keywordBulan)
@@ -66,34 +114,8 @@ class LaporanController extends Controller
                 ->when($request->keywordTahun != null, function ($query) use ($request) {
                     $query->whereYear('riwa_tanggal', $request->keywordTahun)
                 ;})
-                ->orderBy('riwa_tanggal')
+                ->orderBy('riwa_tanggal', 'desc')
                 ->simplePaginate(10);       
-        }
-        elseif ($user->role == 3){
-            $result = Laporan::where('lap_asal', $user->domisili)
-                ->when($request->keywordDaerah != null, function ($query) use ($request) {
-                    $query->where('lap_domisili', $request->keywordDaerah)
-                ;})
-                ->when($request->keywordBulan != null, function ($query) use ($request) {
-                    $query->whereMonth('lap_tanggal', $request->keywordBulan)
-                ;})
-                ->when($request->keywordTahun != null, function ($query) use ($request) {
-                    $query->whereYear('lap_tanggal', $request->keywordTahun)
-                ;})
-                ->get();
-
-            $result2 = Laporan::where('lap_asal', $user->domisili)
-                ->when($request->keywordDaerah != null, function ($query) use ($request) {
-                    $query->where('lap_domisili', $request->keywordDaerah)
-                ;})
-                ->when($request->keywordBulan != null, function ($query) use ($request) {
-                    $query->whereMonth('lap_tanggal', $request->keywordBulan)
-                ;})
-                ->when($request->keywordTahun != null, function ($query) use ($request) {
-                    $query->whereYear('lap_tanggal', $request->keywordTahun)
-                ;})
-                ->orderBy('lap_tanggal')
-                ->simplePaginate(10);
         }
         else {
             $result = Laporan::when($request->keywordDaerah != null, function ($query) use ($request) {
@@ -116,12 +138,12 @@ class LaporanController extends Controller
                 ->when($request->keywordTahun != null, function ($query) use ($request) {
                     $query->whereYear('lap_tanggal', $request->keywordTahun)
                 ;})
-                ->orderBy('lap_tanggal')
+                ->orderBy('lap_tanggal', 'desc')
                 ->simplePaginate(10);
         }
 
-        //  dd($result2->toArray());
-        return view('laporan', compact('result', 'result2', 'query', 'entry'));
+        // dd($result2->toArray());
+        return view('show.laporan', compact('result', 'result2', 'query', 'entry'));
         // return response()->json($result);
     }
 
@@ -133,6 +155,7 @@ class LaporanController extends Controller
     public function create(Request $request)
     {
         $inputTgl = $request->inputTgl;
+        $inputKeg = $request->inputKeg;
         $inputNama = $request->inputNama;
         $inputKab = $request->inputKab;
         $inputDona = $request->inputDona;
@@ -141,6 +164,7 @@ class LaporanController extends Controller
 
         $entry = [
             'inputTgl' =>$inputTgl,
+            'inputKeg' =>$inputKeg,
             'inputNama' =>$inputNama,
             'inputKab' =>$inputKab,
             'inputDona' =>$inputDona,
@@ -148,7 +172,7 @@ class LaporanController extends Controller
             'inputBank' =>$inputBank,
         ];
 
-        return view('inputdonasi', compact('entry'));
+        return view('post.inputdonasi', compact('entry'));
     }
 
     /**
@@ -165,8 +189,14 @@ class LaporanController extends Controller
 
         $laporan->lap_id = $nocek;
         $laporan->lap_tanggal = $request->inputTgl;
+        $laporan->lap_kegiatan = $request->inputKeg;
         $laporan->lap_penerima = $user->nama;
-        $laporan->lap_domisili = $user->domisili;
+        if ($user->id == 2){
+            $laporan->lap_domisili = $request->inputKab;
+        }
+        else {
+            $laporan->lap_domisili = $user->domisili;
+        }
         $laporan->lap_pemberi = $request->inputNama;
         $laporan->lap_asal = $request->inputKab;
         $laporan->lap_jml = $request->inputDona;
@@ -223,7 +253,13 @@ class LaporanController extends Controller
      */
     public function destroy($id)
     {
-        Laporan::where('lap_id', $id)->delete();
+        if (Laporan::where('lap_id', $id)->first()){
+            // It exists
+            Laporan::where('lap_id', $id)->delete();
+        }
+        else{
+            // It does not exist
+        }
 
         if (Riwayat::where('riwa_id', $id)->first()){
             // It exists
@@ -233,8 +269,19 @@ class LaporanController extends Controller
             // It does not exist
         }
 
-        return redirect()->action(
-            'HomeController@index'
-        );
+        return redirect()->back()->with('success', 'Data berhasil dihapus');
+    }
+
+    /**
+     * @return BinaryFileResponse
+     */
+    public function export()
+    {
+        $user = Auth::user();
+
+        $nama = $user->nama;
+        $daerah = $user->domisili;
+
+        return new LaporanExport($nama, $daerah);
     }
 }
